implemented getDevice

+++ include/devicemanager.h
@@ -1,6 +1,7 @@
 /**
  * @file devicemanager.h
  * -----------------------------
+ * @brief This class is a wrapper for the hidapi functionality.
  * @author Bartolone, Kai
  * @date November 2020
  */
@@ -17,26 +18,26 @@
 
 class DeviceManager
 {
-  public:
-    DeviceManager();
-    ~DeviceManager();
+public:
+  DeviceManager();
+  ~DeviceManager();
 
-    /**
-     * @fn listDevices
-     * -----------------
-     * @description: Prints out hid device information (Manufacturer, Product, Release, Interface,
-     * and Usage).
-     */
-    void listDevices();
+  /**
+   * @fn listDevices
+   * -----------------
+   * @description: Prints out hid device information (Manufacturer, Product, Release, Interface,
+   * and Usage).
+   */
+  void listDevices();
 
-    /**
-     * @fn getDevice
-     * -----------------
-     * @description: Searches for hid device with product id or product type.
-     * @returns: first device with matching product id or product type
-     */
-    hid_device* getDevice(int id);
-    hid_device* getDevice(std::string product);
+  /**
+   * @fn getDevice
+   * -----------------
+   * @description: Searches for hid device with product id or product type.
+   * @returns: first device with matching product id or product type
+   */
+  hid_device *getDevice(unsigned short id);
+  hid_device *getDevice(std::string product);
 };
 
 #endif // DEVICEMANAGER_H
